Replace manual SettingsViewModel wiring with data-driven dispatch
- Add read/write lambdas to all non-Custom SettingDef subclasses
- Wire all 28 setting definitions with LlmHttpPrefs read/write lambdas
- Auto-generate SettingEntry map from SettingDef metadata in ViewModel
- Replace 22-case getToggleEntry when block with generic map lookup
- Replace per-key persistSetting when blocks with generic def.write()
- Unify NumericWithUnit entries to Long (was mixed Int/Long)
- Remove .toLong()/.toInt() conversions in AutoLaunchCard/UpdatesCard
- Fix pre-existing bug: bearerTokenEntry not reset in resetToDefaults

+++ Android/src/app/src/main/java/com/ollitert/llm/server/ui/server/settings/AutoLaunchCard.kt
@@ -168,9 +168,9 @@ internal fun AutoLaunchCard(vm: SettingsViewModel, downloadedModelNames: List<St
 
       NumericWithUnitRow(
         def = KEEP_ALIVE_TIMEOUT,
-        baseValue = vm.keepAliveMinutesEntry.current.toLong(),
-        savedBaseValue = vm.keepAliveMinutesEntry.saved.toLong(),
-        onBaseValueChange = { vm.keepAliveMinutesEntry.update(it.toInt()) },
+        baseValue = vm.keepAliveMinutesEntry.current,
+        savedBaseValue = vm.keepAliveMinutesEntry.saved,
+        onBaseValueChange = { vm.keepAliveMinutesEntry.update(it) },
         searchQuery = vm.searchQuery,
         isError = vm.hasError(KEEP_ALIVE_TIMEOUT.key),
         enabled = vm.keepAliveEnabledEntry.current,
